+++ frontend/locale/en/messages.json
@@ -1,7 +1,9 @@
 {
   "ALPHA": "ALPHA",
+  "Amount": "Amount",
   "Call 911": "Call 911",
   "Choose File": "Choose File",
+  "Currency": "Currency",
   "Description": "Description",
   "Email addresses": "Email addresses",
   "Error uploading file": "Error uploading file",
@@ -13,6 +15,7 @@
   "How did that happen?": "How did that happen?",
   "How do you want to make your report?": "How do you want to make your report?",
   "How urgent is your situation?": "How urgent is your situation?",
+  "How were you contacted?": "How were you contacted?",
   "I don’t expect to receive a follow-up": "I don’t expect to receive a follow-up",
   "I have a story to tell about what happened": "I have a story to tell about what happened",
   "I have evidence that could help identify who is responsible": "I have evidence that could help identify who is responsible",
@@ -22,16 +25,22 @@
   "It's an emergency: I need help now because someone's life is in danger.": "It's an emergency: I need help now because someone's life is in danger.",
   "I’d like to report, but it’s not an emergency.": "I’d like to report, but it’s not an emergency.",
   "I’m not sure": "I’m not sure",
+  "Method of payment": "Method of payment",
   "Next": "Next",
   "No files currently selected for upload": "No files currently selected for upload",
+  "Other flow": "Other flow",
+  "Other method of contact": "Other method of contact",
+  "Other method of payment": "Other method of payment",
   "Page not found": "Page not found",
   "Phone numbers": "Phone numbers",
   "Please complete the form to tell us what was affected.": "Please complete the form to tell us what was affected.",
+  "Please fill out all fields": "Please fill out all fields",
   "Privacy": "Privacy",
   "Report a scam": "Report a scam",
   "Report a scam or an online crime": "Report a scam or an online crime",
   "Reports per day": "Reports per day",
   "Reports similar to yours": "Reports similar to yours",
+  "Required": "Required",
   "Sorry, the page you were trying to view does not exist.": "Sorry, the page you were trying to view does not exist.",
   "Submit scam": "Submit scam",
   "Submit!": "Submit!",
@@ -40,10 +49,12 @@
   "This site will be changing as we test ideas and learn from people like you.": "This is a prototype for research purposes only.",
   "Upload a suspicious file": "Upload a suspicious file",
   "Upload another file": "Upload another file",
+  "Was money taken?": "Was money taken?",
   "Website": "Website",
   "Website links": "Website links",
   "What information do you have to report?": "What information do you have to report?",
   "What was affected? Choose all that apply.": "What was affected? Choose all that apply.",
+  "When were you contacted?": "When were you contacted?",
   "Would you like to leave contact information for a potential follow-up?": "Would you like to leave contact information for a potential follow-up?",
   "You can call the call centre": "You can call the call centre",
   "You can report online": "You can report online",
