+++ frontend/locale/fr/messages.json
@@ -1,7 +1,9 @@
 {
   "ALPHA": "ALPHA",
+  "Amount": "",
   "Call 911": "Composez le 911",
   "Choose File": "Choisissez Fichier",
+  "Currency": "",
   "Description": "Description",
   "Email addresses": "Adresses courriel",
   "Error uploading file": "Erreur de téléchargement de fichier",
@@ -13,6 +15,7 @@
   "How did that happen?": "Comment cela s'est-il passé?",
   "How do you want to make your report?": "Comment aimeriez-vous signaler ce méfait?",
   "How urgent is your situation?": "Quelle est l'urgence de votre situation?",
+  "How were you contacted?": "",
   "I don’t expect to receive a follow-up": "Je ne m’attends pas à recevoir de suivi.",
   "I have a story to tell about what happened": "J’ai une expérience à raconter sur ce qui s’est passé",
   "I have evidence that could help identify who is responsible": "J’ai des preuves qui pourraient aider à identifier la personne responsable",
@@ -22,16 +25,22 @@
   "It's an emergency: I need help now because someone's life is in danger.": "Il s’agit d’une urgence. J’ai besoin d’aide maintenant, car la vie de quelqu’un est en danger.",
   "I’d like to report, but it’s not an emergency.": "J’aimerais signaler quelque chose, mais ce n’est pas une urgence.",
   "I’m not sure": "Je ne suis pas certain",
+  "Method of payment": "",
   "Next": "Suivant",
   "No files currently selected for upload": "Aucun fichier actuellement sélectionné pour le téléchargement",
+  "Other flow": "",
+  "Other method of contact": "",
+  "Other method of payment": "",
   "Page not found": "Page introuvable",
   "Phone numbers": "Numéros de téléphone",
   "Please complete the form to tell us what was affected.": "Veuillez remplir le formulaire pour nous dire ce qui a été touché.",
+  "Please fill out all fields": "",
   "Privacy": "Confidentialité",
   "Report a scam": "Signalez une fraude",
   "Report a scam or an online crime": "Signalez une fraude ou un crime informatique",
   "Reports per day": "Alertes par jour",
   "Reports similar to yours": "Alertes semblables à la vôtre",
+  "Required": "",
   "Sorry, the page you were trying to view does not exist.": "Désolé, la page que vous essayiez de voir n'existe pas.",
   "Submit scam": "Soumettre une fraude",
   "Submit!": "Soumettre!",
@@ -40,10 +49,12 @@
   "This site will be changing as we test ideas and learn from people like you.": "Ceci est un prototype à des fins de recherche seulement.",
   "Upload a suspicious file": "Télécharger un fichier suspect",
   "Upload another file": "Télécharger un autre fichier",
+  "Was money taken?": "",
   "Website": "Site web",
   "Website links": "adresse de sites web",
   "What information do you have to report?": "Quel type de renseignements avez-vous à signaler?",
   "What was affected? Choose all that apply.": "Qu'est-ce qui a été affecté? Cochez toutes les réponses qui s'appliquent.",
+  "When were you contacted?": "",
   "Would you like to leave contact information for a potential follow-up?": "Aimeriez-vous laisser vos coordonnées pour un suivi éventuel?",
   "You can call the call centre": "Vous pouvez appeler le centre d'appels",
   "You can report online": "Vous pouvez signaler le méfait en ligne",
